+++ ipfx/x_to_nwb/DatConverter.py
@@ -296,7 +296,7 @@ class DatConverter:
 
                         timeSeries = seriesClass(name=name,
                                                  data=data,
-                                                 sweep_number=cycle_id,
+                                                 sweep_number=np.uint64(cycle_id),
                                                  unit=unit,
                                                  electrode=electrode,
                                                  gain=gain,
@@ -390,7 +390,7 @@ class DatConverter:
 
                             acquistion_data = seriesClass(name=name,
                                                           data=data,
-                                                          sweep_number=cycle_id,
+                                                          sweep_number=np.uint64(cycle_id),
                                                           unit=unit,
                                                           electrode=electrode,
                                                           gain=gain,
@@ -423,7 +423,7 @@ class DatConverter:
 
                             acquistion_data = seriesClass(name=name,
                                                           data=data,
-                                                          sweep_number=cycle_id,
+                                                          sweep_number=np.uint64(cycle_id),
                                                           unit=unit,
                                                           electrode=electrode,
                                                           gain=gain,
